Incompatible HTML changes
Trying to get more HTML5 elements used.

PrintAllPages:

Use the article element instead of a div with class "page". The new
article element still has the "h-entry" class that the old div had.

The h1 element for these pages used to have the class "entry-title"
which is apparently deprecated. The new code now uses the "p-name"
attribute.

The page content is no longer surrounded in a div with the
"entry-content" class and the appropriate lang attribute. We rely on
PrintPageHtml to do the right thing, now.

PrintPageHtml:

Surround the page being printed with a div containing the "e-content"
class and an appropriate lang attribute.

PageHtml:

This also uses PrintPageHtml and therefore doesn't need to surround
the page content with a div containing the "page" class and the lang
attribute.

As PageHtml is used in RSS feed generation, that means that the feed
entries now don't have a div containing the "page" class but a div
containing the "e-content" class.

GetHeaderDiv:

Instead of using a div with the "header" class, use the header
element.

Instead of using a div with the "menu" class, use the nav element.

PrintPageContent:

No changes! We're not changing the div here because the content that
is being printed here does not belong into an article element. It is
not "a self-contained composition in a … page … intended to be
independently distributable or reusable" – it *is* the page
itself (without the h1 header).

PrintFooter:

Use an additional footer element.

DefaultFooter:

Remove the div with the "footer" class.

References:

* http://microformats.org/wiki/h-entry
* https://developer.mozilla.org/en-US/docs/Web/HTML/Element/header
* https://developer.mozilla.org/en-US/docs/Web/HTML/Element/nav
* https://developer.mozilla.org/en-US/docs/Web/HTML/Element/article
* https://developer.mozilla.org/en-US/docs/Web/HTML/Element/footer